penambahan function verifysensorkey dan listsensors
