+++ app/function/getauthtoken.py
@@ -6,8 +6,6 @@ from flask_httpauth import HTTPBasicAuth
 from datetime import datetime
 import os, tarfile
 
-@app.route('/api/token/v1.0/getauthtoken', methods=['POST'])
-@auth.login_required
 def getauthtoken():
     token = g.user.generate_auth_token()
     return jsonify({'token': token.decode('ascii')})
